Moving execute to utils.py
